+++ arrays/src/test/java/by/tolkach/arrays/ArrayAnalyzerTest.java
@@ -15,7 +15,7 @@ public class ArrayAnalyzerTest {
 
     @ParameterizedTest
     @MethodSource("arrayProvider")
-    void analyze(long[] expectedArray, long[] providedArray) {
+    void analyze(Integer[] expectedArray, Integer[] providedArray) {
         arrayAnalyzer = new ArrayCycleAnalyser();
         Assertions.assertArrayEquals(expectedArray, arrayAnalyzer.analyze(providedArray));
         arrayAnalyzer = new ArrayStreamAnalyzer();
@@ -24,10 +24,10 @@ public class ArrayAnalyzerTest {
 
     static Stream<Arguments> arrayProvider() {
         return Stream.of(
-                arguments(new long[] {10, -65}, new long[] {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, -11, -12, -13, -14, -15}),
-                arguments(new long[] {8, -50}, new long[] {0, 2, 3, 0, 5, 6, 7, 8, 9, 10, -11, -12, -13, -14}),
-                arguments(new long[0], new long[0]),
-                arguments(new long[0], null)
+                arguments(new Integer[] {10, -65}, new Integer[] {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, -11, -12, -13, -14, -15}),
+                arguments(new Integer[] {8, -50}, new Integer[] {0, 2, 3, 0, 5, 6, 7, 8, 9, 10, -11, -12, -13, -14}),
+                arguments(new Integer[0], new Integer[0]),
+                arguments(new Integer[0], null)
         );
     }
 }
